tkf-data-layer/pyproject.toml
[tool.poetry]
name = "tkf-data-layer"
version = "0.1.0"
description = "TKF data layer: Neo4j-backed graph API for evidence-gathering-agent"
authors = ["TKF Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = { version = "^0.27", extras = ["standard"] }
pydantic = "^2.5"
python-dotenv = "^1.0.1"
neo4j = "^5.16"
sentence-transformers = "^3.0.0"
numpy = ">=1.26,<3"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
